frontend: router: Extract getRouteUseClusterURL to own file
