+++ frontend/src/lib/router/getRouteUseClusterURL.tsx
@@ -1,0 +1,52 @@
+/*
+ * Copyright 2025 The Kubernetes Authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+/*
+ * Copyright 2025 The Kubernetes Authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import type { Route } from './Route';
+
+/**
+ * Should the route use a cluster URL?
+ *
+ * @param route
+ * @returns true when a cluster URL contains cluster in the URL. eg. /c/minikube/my-url
+ *   false, the URL does not contain the cluster. eg. /my-url
+ */
+export function getRouteUseClusterURL(route: Route): boolean {
+  if (route.useClusterURL === undefined && route.noCluster !== undefined) {
+    console.warn('Route.noCluster is deprecated. Please use route.useClusterURL instead.');
+    return route.noCluster;
+  }
+  if (route.useClusterURL === undefined) {
+    // default is true, so undefined === true.
+    return true;
+  }
+  return route.useClusterURL;
+}
