+++ frontend/src/lib/router/index.tsx
@@ -125,9 +125,10 @@ import ReplicaSet from '../k8s/replicaSet';
 import StatefulSet from '../k8s/statefulSet';
 import { getClusterPrefixedPath } from '../util';
 import { getDefaultRoutes, setDefaultRoutes } from './getDefaultRoutes';
+import { getRouteUseClusterURL } from './getRouteUseClusterURL';
 import type { Route } from './Route';
 
-export { getDefaultRoutes };
+export { getDefaultRoutes, getRouteUseClusterURL };
 export type { Route };
 
 const LazyGraphView = React.lazy(() =>
@@ -984,25 +985,6 @@ export function getRoute(routeName: string) {
     }
   }
   return getDefaultRoutes()[routeKey];
-}
-
-/**
- * Should the route use a cluster URL?
- *
- * @param route
- * @returns true when a cluster URL contains cluster in the URL. eg. /c/minikube/my-url
- *   false, the URL does not contain the cluster. eg. /my-url
- */
-export function getRouteUseClusterURL(route: Route): boolean {
-  if (route.useClusterURL === undefined && route.noCluster !== undefined) {
-    console.warn('Route.noCluster is deprecated. Please use route.useClusterURL instead.');
-    return route.noCluster;
-  }
-  if (route.useClusterURL === undefined) {
-    // default is true, so undefined === true.
-    return true;
-  }
-  return route.useClusterURL;
 }
 
 export function getRoutePath(route: Route) {
